Use esbuild for creating the custom.js file
