+++ .github/workflows/create-customization-package-zip.yml
@@ -8,17 +8,18 @@ jobs:
           - name: Checkout
             uses: actions/checkout@v3
             
-          - name: Use Node.js 16
+          - name: Use Node.js 20
             uses: actions/setup-node@v3
             with:
-              node-version: 16.x
+              node-version: 20.x
 
           - name: Build
             run: |
               npm install
               gulp cleanCSS
               gulp styles
-
+              node build.js 
+              
           - name: Archive production artifacts
             uses: actions/upload-artifact@v3
             with:
