+++ src/components/prmLocationItemsAfter.js
@@ -1,0 +1,9 @@
+export const prmLocationItemsAfter = {
+  bindings: { parentCtrl: '<' },
+  templateUrl: "/discovery/custom/01JHU_INST-JHU/html/prm-location-items-after.html",
+  controller: function () {
+    this.$onInit = function () {
+      console.log("prmLocationItemsAfter");
+    };
+  }
+};
