+++ src/components/prmSearchBookmarkFilterAfter.js
@@ -1,0 +1,3 @@
+export const prmSearchBookmarkFilterAfter = {
+  templateUrl: "/discovery/custom/01JHU_INST-JHU/html/prm-search-bookmark-filter-after.html"
+};
