+++ src/components/prmSearchResultThumbnailContainerAfter.js
@@ -1,0 +1,4 @@
+export const prmSearchResultThumbnailContainerAfter = {
+  bindings: { parentCtrl: `<` },
+  templateUrl: "/discovery/custom/01JHU_INST-JHU/html/prm-search-result-thumbnail-container-after.html"
+};
