+++ src/components/prmTopBarBefore.js
@@ -1,0 +1,4 @@
+export const prmTopBarBefore = {
+  bindings: { parentCtrl: `<` },
+  templateUrl: "/discovery/custom/01JHU_INST-JHU/html/prm-top-bar-before.html"
+};
