+++ src/interceptors/config.js
@@ -1,0 +1,5 @@
+export function configureInterceptors(app) {
+  app.config(['$httpProvider', function ($httpProvider) {
+    $httpProvider.interceptors.push('itemRequestUrlInterceptor');
+  }]);
+}
