+++ src/services/AuthService.js
@@ -1,0 +1,12 @@
+export function AuthService() {
+  var isLoggedIn = false;
+
+  return {
+    setIsLoggedIn: function (value) {
+      isLoggedIn = value !== false;
+    },
+    getIsLoggedIn: function () {
+      return isLoggedIn;
+    }
+  };
+}
