+++ src/services/CapitalizeService.js
@@ -1,0 +1,6 @@
+export function CapitalizeService() {
+  this.capitalize = function (string) {
+    if (!string) return string;
+    return string.charAt(0).toUpperCase() + string.slice(1);
+  }
+}
